Extract API Logic into a Service
