+++ arabic-learning-app/src/app/page.tsx
@@ -6,6 +6,7 @@ import Link from "next/link";
 import { Recipe } from "./types/recipe";
 import { sanitizeJsonString } from "./utils/sanitizeJsonString";
 import { STORAGE_KEY, MAX_SAVED_RECIPES } from "./constants";
+import { searchRecipe } from "./services/recipeService";
 
 export default function Home() {
   const [searchedRecipes, setSearchedRecipes] = useState<Recipe[]>([]);
@@ -157,6 +158,15 @@ export default function Home() {
     },
   ]);
 
+  const handleSearch = async (query: string) => {
+    try {
+      const data = await searchRecipe(query);
+      return data;
+    } catch (error) {
+      console.error("Error searching recipe:", error);
+    }
+  };
+
   return (
     <div className="min-h-screen bg-gray-100 text-green-900">
       {/* Hero Section */}
@@ -186,82 +196,9 @@ export default function Home() {
                   setIsSearching(true);
 
                   try {
-                    const requestBody = {
-                      model: "llama3",
-                      prompt: `Generate a valid JSON object for a Middle Eastern ${query} recipe. The JSON must strictly follow this structure:
-                      {
-                      "title": "Recipe name in English",
-                      "arabicTitle": "Recipe name in Arabic",
-                      "transliteration": "Arabic transliteration of the title",
-                      "description": "Brief description of the dish",
-                      "question": {
-                        "arabic": "How to make this dish? (in Arabic)",
-                        "transliteration": "Transliteration of the question",
-                        "english": "How to make this dish?"
-                      },
-                      "ingredients": [
-                        {
-                          "name": "Ingredient name in English",
-                          "arabicName": "Ingredient name in Arabic",
-                          "transliteration": "Arabic transliteration of the ingredient name",
-                          "image": "/ingredient-image.jpg",
-                          "description": "Brief description of the ingredient"
-                        }
-                      ],
-                      "recipeEnglish": "Step-by-step instructions in English.",
-                      "recipeArabic": "Step-by-step instructions in Arabic"
-                    }
-                    Requirements:
-                    * Each recipe step must be in a list format, ensuring they appear on separate lines.
-                    * Ensure all Arabic text is properly written.
-                    * Include accurate transliterations.
-                    * Provide clear and structured ingredient details.
-                    * Return the response ONLY as a JSON object with no additional text, explanations, or formatting.
-                      `,
-                      stream: false,
-                    };
-
-                    const response = await fetch(
-                      "http://localhost:11434/api/generate",
-                      {
-                        method: "POST",
-                        headers: {
-                          "Content-Type": "application/json",
-                        },
-                        body: JSON.stringify(requestBody),
-                      }
-                    );
-
-                    console.log("Response:--------", response);
-
-                    if (!response.ok) {
-                      throw new Error(`HTTP error! status: ${response.status}`);
-                    }
-
-                    const data = await response.json();
-                    if (!data.response) {
-                      throw new Error(
-                        "No recipe data received. Please try again."
-                      );
-                    }
-
-                    let jsonStr = data.response.trim();
-                    console.log("Raw JSON string:", jsonStr);
-
-                    // Extract JSON if it's wrapped in backticks or other text
-                    const jsonMatch = jsonStr.match(/({[\s\S]*})/);
-                    if (jsonMatch) {
-                      jsonStr = jsonMatch[0];
-                    }
-
-                    if (!jsonStr.startsWith("{") || !jsonStr.endsWith("}")) {
-                      throw new Error(
-                        "Invalid recipe data received. Please try again."
-                      );
-                    }
-
                     // Sanitize the JSON string
-                    const sanitizedJson = sanitizeJsonString(jsonStr);
+                    const searchResult = await handleSearch(query);
+                    const sanitizedJson = sanitizeJsonString(searchResult);
 
                     try {
                       const recipeData = JSON.parse(sanitizedJson) as Recipe;
@@ -297,80 +234,9 @@ export default function Home() {
                 setSearchError(null);
 
                 try {
-                  const requestBody = {
-                    model: "llama3",
-                    prompt: `Generate a valid JSON object for a Middle Eastern ${query} recipe. The JSON must strictly follow this structure:
-                      {
-                      "title": "Recipe name in English",
-                      "arabicTitle": "Recipe name in Arabic",
-                      "transliteration": "Arabic transliteration of the title",
-                      "description": "Brief description of the dish",
-                      "question": {
-                        "arabic": "How to make this dish? (in Arabic)",
-                        "transliteration": "Transliteration of the question",
-                        "english": "How to make this dish?"
-                      },
-                      "ingredients": [
-                        {
-                          "name": "Ingredient name in English",
-                          "arabicName": "Ingredient name in Arabic",
-                          "transliteration": "Arabic transliteration of the ingredient name",
-                          "image": "/ingredient-image.jpg",
-                          "description": "Brief description of the ingredient"
-                        }
-                      ],
-                    "recipeEnglish": "Step-by-step instructions in English.",
-                    "recipeArabic": "Step-by-step instructions in Arabic"
-                    }
-                    Requirements:
-                    * Each recipe step must be in a list format, ensuring they appear on separate lines.
-                    * Ensure all Arabic text is properly written.
-                    * Include accurate transliterations.
-                    * Provide clear and structured ingredient details.
-                    * Return the response ONLY as a JSON object with no additional text, explanations, or formatting.
-                      `,
-
-                    stream: false,
-                  };
-
-                  const response = await fetch(
-                    "http://localhost:11434/api/generate",
-                    {
-                      method: "POST",
-                      headers: {
-                        "Content-Type": "application/json",
-                      },
-                      body: JSON.stringify(requestBody),
-                    }
-                  );
-
-                  if (!response.ok) {
-                    throw new Error(`HTTP error! status: ${response.status}`);
-                  }
-
-                  const data = await response.json();
-                  if (!data.response) {
-                    throw new Error(
-                      "No recipe data received. Please try again."
-                    );
-                  }
-
-                  let jsonStr = data.response.trim();
-
-                  // Extract JSON if it's wrapped in backticks or other text
-                  const jsonMatch = jsonStr.match(/({[\s\S]*})/);
-                  if (jsonMatch) {
-                    jsonStr = jsonMatch[0];
-                  }
-
-                  if (!jsonStr.startsWith("{") || !jsonStr.endsWith("}")) {
-                    throw new Error(
-                      "Invalid recipe data received. Please try again."
-                    );
-                  }
-
+                  const showSearchResult = await handleSearch(query);
                   // Sanitize the JSON string
-                  const sanitizedJson = sanitizeJsonString(jsonStr);
+                  const sanitizedJson = sanitizeJsonString(showSearchResult);
 
                   try {
                     const recipeData: Recipe = {
